+++ packages/malloy/src/lang/malloy-parse-info.ts
@@ -21,7 +21,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import {ParseInfo} from './run-malloy-parser';
+import {ParseInfo} from './utils';
 
 export interface MalloyParseInfo extends ParseInfo {
   importBaseURL: string;
